Replace media query with display and conditionally render links
